+++ models/discount/enum/type.go
@@ -1,0 +1,8 @@
+package enum
+
+type Type string
+
+const (
+	TypeFixedAmount Type = "fixed_amount"
+	TypePercentage  Type = "percentage"
+)