+++ models/inappsubscription/enum/store_status.go
@@ -1,0 +1,9 @@
+package enum
+
+type StoreStatus string
+
+const (
+	StoreStatusInTrial   StoreStatus = "in_trial"
+	StoreStatusActive    StoreStatus = "active"
+	StoreStatusCancelled StoreStatus = "cancelled"
+)